requirements.txt
aiofiles==24.1.0
aiohttp==3.9.1
aiosqlite==0.21.0
protobuf==6.32.1
pydantic==1.10.12
pytest==7.4.3
pytest_asyncio==0.21.1
python-dotenv==1.1.1
PyYAML==6.0.2
rich==14.1.0
